requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
python-socketio==5.14.1
python-engineio==4.12.3
eventlet==0.33.3
python-dotenv==1.0.1
gunicorn==21.2.0
google-generativeai
psycopg2-binary==2.9.7
